requirements.txt
aiofiles==23.1.0
aiogram==3.0.0
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
black==23.10.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
frozenlist==1.4.0
idna==3.4
magic-filter==1.0.12
MarkupSafe==2.1.3
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
PyYAML==6.0.1
typing_extensions==4.7.1
yarl==1.9.2
